+++ Dynavity/Dynavity/persistence/data-mapping/UIImageDTO.swift
@@ -12,3 +12,5 @@ struct UIImageDTO: Mappable {
         return image
     }
 }
+
+extension UIImageDTO: Codable {}
